+++ libraries/FRAM_I2C/FRAM.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: FRAM.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.5.3
+// VERSION: 0.5.4
 //    DATE: 2018-01-24
 // PURPOSE: Arduino library for I2C FRAM
 //     URL: https://github.com/RobTillaart/FRAM_I2C
@@ -89,115 +89,130 @@ bool FRAM::isConnected()
 }
 
 
-void FRAM::write8(uint16_t memaddr, uint8_t value)
+void FRAM::write8(uint16_t memAddr, uint8_t value)
 {
   uint8_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint8_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint8_t));
 }
 
 
-void FRAM::write16(uint16_t memaddr, uint16_t value)
+void FRAM::write16(uint16_t memAddr, uint16_t value)
 {
   uint16_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint16_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint16_t));
 }
 
 
-void FRAM::write32(uint16_t memaddr, uint32_t value)
+void FRAM::write32(uint16_t memAddr, uint32_t value)
 {
   uint32_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint32_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint32_t));
 }
 
 
-void FRAM::writeFloat(uint16_t memaddr, float value)
+void FRAM::write64(uint16_t memAddr, uint64_t value)
+{
+  uint64_t val = value;
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint64_t));
+}
+
+
+void FRAM::writeFloat(uint16_t memAddr, float value)
 {
   float val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(float));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(float));
 }
 
 
-void FRAM::writeDouble(uint16_t memaddr, double value)
+void FRAM::writeDouble(uint16_t memAddr, double value)
 {
   double val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(double));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(double));
 }
 
 
-void FRAM::write(uint16_t memaddr, uint8_t * obj, uint16_t size)
+void FRAM::write(uint16_t memAddr, uint8_t * obj, uint16_t size)
 {
   const int blocksize = 24;
   uint8_t * p = obj;
   while (size >= blocksize)
   {
-    _writeBlock(memaddr, p, blocksize);
-    memaddr += blocksize;
+    _writeBlock(memAddr, p, blocksize);
+    memAddr += blocksize;
     p += blocksize;
     size -= blocksize;
   }
   //  remaining
   if (size > 0)
   {
-    _writeBlock(memaddr, p, size);
+    _writeBlock(memAddr, p, size);
   }
 }
 
 
-uint8_t FRAM::read8(uint16_t memaddr)
+uint8_t FRAM::read8(uint16_t memAddr)
 {
   uint8_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint8_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint8_t));
   return val;
 }
 
 
-uint16_t FRAM::read16(uint16_t memaddr)
+uint16_t FRAM::read16(uint16_t memAddr)
 {
   uint16_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint16_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint16_t));
   return val;
 }
 
 
-uint32_t FRAM::read32(uint16_t memaddr)
+uint32_t FRAM::read32(uint16_t memAddr)
 {
   uint32_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint32_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint32_t));
   return val;
 }
 
 
-float FRAM::readFloat(uint16_t memaddr)
+uint64_t FRAM::read64(uint16_t memAddr)
+{
+  uint64_t val;
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint64_t));
+  return val;
+}
+
+
+float FRAM::readFloat(uint16_t memAddr)
 {
   float val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(float));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(float));
   return val;
 }
 
 
-double FRAM::readDouble(uint16_t memaddr)
+double FRAM::readDouble(uint16_t memAddr)
 {
   double val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(double));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(double));
   return val;
 }
 
 
-void FRAM::read(uint16_t memaddr, uint8_t * obj, uint16_t size)
+void FRAM::read(uint16_t memAddr, uint8_t * obj, uint16_t size)
 {
   const uint8_t blocksize = 24;
   uint8_t * p = obj;
   while (size >= blocksize)
   {
-    _readBlock(memaddr, p, blocksize);
-    memaddr += blocksize;
+    _readBlock(memAddr, p, blocksize);
+    memAddr += blocksize;
     p += blocksize;
     size -= blocksize;
   }
   // remainder
   if (size > 0)
   {
-    _readBlock(memaddr, p, size);
+    _readBlock(memAddr, p, size);
   }
 }
 
@@ -205,15 +220,15 @@ void FRAM::read(uint16_t memaddr, uint8_t * obj, uint16_t size)
 ////////////////////////////////////////////////////////////////////////
 
 
-int32_t FRAM::readUntil(uint16_t memaddr, char * buf, uint16_t buflen, char separator)
+int32_t FRAM::readUntil(uint16_t memAddr, char * buffer, uint16_t bufferLength, char separator)
 {
   //  read and fill the buffer at once.
-  read(memaddr, (uint8_t *)buf, buflen);
-  for (uint16_t length = 0; length < buflen; length++)
+  read(memAddr, (uint8_t *)buffer, bufferLength);
+  for (uint16_t length = 0; length < bufferLength; length++)
   {
-    if (buf[length] == separator)
+    if (buffer[length] == separator)
     {
-      buf[length] = 0;    //  replace separator => \0 EndChar
+      buffer[length] = 0;    //  replace separator => \0 EndChar
       return length;
     }
   }
@@ -222,15 +237,15 @@ int32_t FRAM::readUntil(uint16_t memaddr, char * buf, uint16_t buflen, char sepa
 }
 
 
-int32_t FRAM::readLine(uint16_t memaddr, char * buf, uint16_t buflen)
+int32_t FRAM::readLine(uint16_t memAddr, char * buffer, uint16_t bufferLength)
 {
   //  read and fill the buffer at once.
-  read(memaddr, (uint8_t *)buf, buflen);
-  for (uint16_t length = 0; length < buflen-1; length++)
+  read(memAddr, (uint8_t *)buffer, bufferLength);
+  for (uint16_t length = 0; length < bufferLength-1; length++)
   {
-    if (buf[length] == '\n')
+    if (buffer[length] == '\n')
     {
-      buf[length + 1] = 0;    //  add \0 EndChar after '\n'
+      buffer[length + 1] = 0;    //  add \0 EndChar after '\n'
       return length + 1;
     }
   }
@@ -259,22 +274,32 @@ bool FRAM::getWriteProtect()
 
 uint16_t FRAM::getManufacturerID()
 {
-  return _getMetaData(0);
+  uint32_t value = _getMetaData();
+  return (value >> 12) & 0x0FFF;
 }
 
 
 uint16_t FRAM::getProductID()
 {
-  return _getMetaData(1);
+  uint32_t value = _getMetaData();
+  return value & 0x0FFF;
 }
 
 
-// NOTE: returns the size in kiloBYTE
+//  DENSITY
+//           Fujitsu data sheet
+//  3 =>     MB85RC64 = 64 Kbit.
+//  5 =>     MB85RC256
+//  6 =>     MB85RC512
+//  7 =>     MB85RC1M
+//  NOTE: returns the size in kiloBYTE (0 is read error)
 uint16_t FRAM::getSize()
 {
-  uint16_t density = _getMetaData(2);
-  uint16_t size = 0;
-  if (density > 0) size = (1UL << density);
+  uint32_t value = _getMetaData();
+  if (value == 0xFFFFFFFF) return 0;
+
+  uint16_t density = (value >> 8) & 0x0F;
+  uint16_t size = (1UL << density);
   _sizeBytes = size * 1024UL;
   return size;
 }
@@ -336,20 +361,18 @@ bool FRAM::wakeup(uint32_t trec)
 // FRAM PROTECTED
 //
 
-//  metadata is packed as  [....MMMM][MMMMDDDD][PPPPPPPP]
+//  metadata is packed as  [MMMMMMMM][MMMMDDDD][PPPPPPPP]
 //  M = manufacturerID
 //  D = density => memory size = 2^D KB
 //  P = product ID (together with D)
 //  P part might be proprietary
-uint16_t FRAM::_getMetaData(uint8_t field)
+uint32_t FRAM::_getMetaData()
 {
-  if (field > 2) return 0;
-
   _wire->beginTransmission(FRAM_SLAVE_ID_);
   _wire->write(_address << 1);
   _wire->endTransmission(false);
   int x = _wire->requestFrom(FRAM_SLAVE_ID_, (uint8_t)3);
-  if (x != 3) return -1;
+  if (x != 3) return 0xFFFFFFFF;
 
   uint32_t value = 0;
   value = _wire->read();
@@ -358,26 +381,15 @@ uint16_t FRAM::_getMetaData(uint8_t field)
   value = value << 8;
   value |= _wire->read();
 
-  //  MANUFACTURER
-  if (field == 0) return (value >> 12) & 0xFF;
-  //  PRODUCT ID
-  if (field == 1) return value & 0x0FFF;
-  //  DENSITY
-  //           Fujitsu data sheet
-  //  3 =>     MB85RC64 = 64 Kbit.
-  //  5 =>     MB85RC256
-  //  6 =>     MB85RC512
-  //  7 =>     MB85RC1M
-  if (field == 2) return (value >> 8) & 0x0F;
-  return 0;
+  return value;
 }
 
 
-void FRAM::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM::_writeBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   _wire->beginTransmission(_address);
-  _wire->write((uint8_t) (memaddr >> 8));
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr >> 8));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   uint8_t * p = obj;
   for (uint8_t i = size; i > 0; i--)
   {
@@ -387,11 +399,11 @@ void FRAM::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
 }
 
 
-void FRAM::_readBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM::_readBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   _wire->beginTransmission(_address);
-  _wire->write((uint8_t) (memaddr >> 8));
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr >> 8));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   _wire->endTransmission();
   _wire->requestFrom(_address, size);
   uint8_t * p = obj;
@@ -412,128 +424,143 @@ FRAM32::FRAM32(TwoWire *wire) : FRAM(wire)
 }
 
 
-void FRAM32::write8(uint32_t memaddr, uint8_t value)
+void FRAM32::write8(uint32_t memAddr, uint8_t value)
 {
   uint8_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint8_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint8_t));
 }
 
 
-void FRAM32::write16(uint32_t memaddr, uint16_t value)
+void FRAM32::write16(uint32_t memAddr, uint16_t value)
 {
   uint16_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint16_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint16_t));
 }
 
 
-void FRAM32::write32(uint32_t memaddr, uint32_t value)
+void FRAM32::write32(uint32_t memAddr, uint32_t value)
 {
   uint32_t val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(uint32_t));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint32_t));
 }
 
 
-void FRAM32::writeFloat(uint32_t memaddr, float value)
+void FRAM32::write64(uint32_t memAddr, uint64_t value)
+{
+  uint64_t val = value;
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(uint64_t));
+}
+
+
+void FRAM32::writeFloat(uint32_t memAddr, float value)
 {
   float val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(float));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(float));
 }
 
 
-void FRAM32::writeDouble(uint32_t memaddr, double value)
+void FRAM32::writeDouble(uint32_t memAddr, double value)
 {
   double val = value;
-  _writeBlock(memaddr, (uint8_t *)&val, sizeof(double));
+  _writeBlock(memAddr, (uint8_t *)&val, sizeof(double));
 }
 
 
-void FRAM32::write(uint32_t memaddr, uint8_t * obj, uint16_t size)
+void FRAM32::write(uint32_t memAddr, uint8_t * obj, uint16_t size)
 {
   const int blocksize = 24;
   uint8_t * p = obj;
   while (size >= blocksize)
   {
-    _writeBlock(memaddr, p, blocksize);
-    memaddr += blocksize;
+    _writeBlock(memAddr, p, blocksize);
+    memAddr += blocksize;
     p += blocksize;
     size -= blocksize;
   }
   // remaining
   if (size > 0)
   {
-    _writeBlock(memaddr, p, size);
+    _writeBlock(memAddr, p, size);
   }
 }
 
 
-uint8_t FRAM32::read8(uint32_t memaddr)
+uint8_t FRAM32::read8(uint32_t memAddr)
 {
   uint8_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint8_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint8_t));
   return val;
 }
 
 
-uint16_t FRAM32::read16(uint32_t memaddr)
+uint16_t FRAM32::read16(uint32_t memAddr)
 {
   uint16_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint16_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint16_t));
   return val;
 }
 
 
-uint32_t FRAM32::read32(uint32_t memaddr)
+uint32_t FRAM32::read32(uint32_t memAddr)
 {
   uint32_t val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(uint32_t));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint32_t));
   return val;
 }
 
 
-float FRAM32::readFloat(uint32_t memaddr)
+uint64_t FRAM32::read64(uint32_t memAddr)
+{
+  uint64_t val;
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(uint64_t));
+  return val;
+}
+
+
+float FRAM32::readFloat(uint32_t memAddr)
 {
   float val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(float));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(float));
   return val;
 }
 
 
-double FRAM32::readDouble(uint32_t memaddr)
+double FRAM32::readDouble(uint32_t memAddr)
 {
   double val;
-  _readBlock(memaddr, (uint8_t *)&val, sizeof(double));
+  _readBlock(memAddr, (uint8_t *)&val, sizeof(double));
   return val;
 }
 
 
-void FRAM32::read(uint32_t memaddr, uint8_t * obj, uint16_t size)
+void FRAM32::read(uint32_t memAddr, uint8_t * obj, uint16_t size)
 {
   const uint8_t blocksize = 24;
   uint8_t * p = obj;
   while (size >= blocksize)
   {
-    _readBlock(memaddr, p, blocksize);
-    memaddr += blocksize;
+    _readBlock(memAddr, p, blocksize);
+    memAddr += blocksize;
     p += blocksize;
     size -= blocksize;
   }
   // remainder
   if (size > 0)
   {
-    _readBlock(memaddr, p, size);
+    _readBlock(memAddr, p, size);
   }
 }
 
 
-int32_t FRAM32::readUntil(uint32_t memaddr, char * buf, uint16_t buflen, char separator)
+int32_t FRAM32::readUntil(uint32_t memAddr, char * buffer, uint16_t bufferLength, char separator)
 {
   //  read and fill the buffer at once.
-  read(memaddr, (uint8_t *)buf, buflen);
-  for (uint16_t length = 0; length < buflen; length++)
+  read(memAddr, (uint8_t *)buffer, bufferLength);
+  for (uint16_t length = 0; length < bufferLength; length++)
   {
-    if (buf[length] == separator)
+    if (buffer[length] == separator)
     {
-      buf[length] = 0;    //  replace separator => \0 EndChar
+      buffer[length] = 0;    //  replace separator => \0 EndChar
       return length;
     }
   }
@@ -542,15 +569,15 @@ int32_t FRAM32::readUntil(uint32_t memaddr, char * buf, uint16_t buflen, char se
 }
 
 
-int32_t FRAM32::readLine(uint32_t memaddr, char * buf, uint16_t buflen)
+int32_t FRAM32::readLine(uint32_t memAddr, char * buffer, uint16_t bufferLength)
 {
   //  read and fill the buffer at once.
-  read(memaddr, (uint8_t *)buf, buflen);
-  for (uint16_t length = 0; length < buflen-1; length++)
+  read(memAddr, (uint8_t *)buffer, bufferLength);
+  for (uint16_t length = 0; length < bufferLength-1; length++)
   {
-    if (buf[length] == '\n')
+    if (buffer[length] == '\n')
     {
-      buf[length + 1] = 0;    //  add \0 EndChar after '\n'
+      buffer[length + 1] = 0;    //  add \0 EndChar after '\n'
       return length + 1;
     }
   }
@@ -559,29 +586,29 @@ int32_t FRAM32::readLine(uint32_t memaddr, char * buf, uint16_t buflen)
 }
 
 
-template <class T> uint32_t FRAM32::writeObject(uint32_t memaddr, T &obj)
+template <class T> uint32_t FRAM32::writeObject(uint32_t memAddr, T &obj)
 {
-  write(memaddr, (uint8_t *) &obj, sizeof(obj));
-  return memaddr + sizeof(obj);
+  write(memAddr, (uint8_t *) &obj, sizeof(obj));
+  return memAddr + sizeof(obj);
 };
 
 
-template <class T> uint32_t FRAM32::readObject(uint32_t memaddr, T &obj)
+template <class T> uint32_t FRAM32::readObject(uint32_t memAddr, T &obj)
 {
-  read(memaddr, (uint8_t *) &obj, sizeof(obj));
-  return memaddr + sizeof(obj);
+  read(memAddr, (uint8_t *) &obj, sizeof(obj));
+  return memAddr + sizeof(obj);
 }
 
 
 uint32_t FRAM32::clear(uint8_t value)
 {
-  uint8_t buf[16];
-  for (uint8_t i = 0; i < 16; i++) buf[i] = value;
+  uint8_t buffer[16];
+  for (uint8_t i = 0; i < 16; i++) buffer[i] = value;
   uint32_t start = 0;
   uint32_t end = _sizeBytes;
   for (uint32_t addr = start; addr < end; addr += 16)
   {
-    _writeBlock(addr, buf, 16);
+    _writeBlock(addr, buffer, 16);
   }
   return end - start;
 }
@@ -593,14 +620,14 @@ uint32_t FRAM32::clear(uint8_t value)
 //
 //  FRAM32  PROTECTED
 //
-void FRAM32::_writeBlock(uint32_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM32::_writeBlock(uint32_t memAddr, uint8_t * obj, uint8_t size)
 {
   uint8_t _addr = _address;
-  if (memaddr & 0x00010000) _addr += 0x01;
+  if (memAddr & 0x00010000) _addr += 0x01;
 
   _wire->beginTransmission(_addr);
-  _wire->write((uint8_t) (memaddr >> 8));
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr >> 8));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   uint8_t * p = obj;
   for (uint8_t i = size; i > 0; i--)
   {
@@ -610,14 +637,14 @@ void FRAM32::_writeBlock(uint32_t memaddr, uint8_t * obj, uint8_t size)
 }
 
 
-void FRAM32::_readBlock(uint32_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM32::_readBlock(uint32_t memAddr, uint8_t * obj, uint8_t size)
 {
   uint8_t _addr = _address;
-  if (memaddr & 0x00010000) _addr += 0x01;
+  if (memAddr & 0x00010000) _addr += 0x01;
 
   _wire->beginTransmission(_address);
-  _wire->write((uint8_t) (memaddr >> 8));
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr >> 8));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   _wire->endTransmission();
   _wire->requestFrom(_addr, size);
   uint8_t * p = obj;
@@ -669,12 +696,12 @@ uint16_t FRAM11::getSize()
 //
 //  FRAM11  PROTECTED
 //
-void FRAM11::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM11::_writeBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   // Device uses Address Pages
-  uint8_t  DeviceAddrWithPageBits = _address | ((memaddr & 0x0700) >> 8);
+  uint8_t  DeviceAddrWithPageBits = _address | ((memAddr & 0x0700) >> 8);
   _wire->beginTransmission(DeviceAddrWithPageBits);
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr & 0xFF));
 
   uint8_t * p = obj;
   for (uint8_t i = size; i > 0; i--)
@@ -685,12 +712,12 @@ void FRAM11::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
 }
 
 
-void FRAM11::_readBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM11::_readBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   // Device uses Address Pages
-  uint8_t DeviceAddrWithPageBits = _address | ((memaddr & 0x0700) >> 8);
+  uint8_t DeviceAddrWithPageBits = _address | ((memAddr & 0x0700) >> 8);
   _wire->beginTransmission(DeviceAddrWithPageBits);
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   _wire->endTransmission();
   _wire->requestFrom(DeviceAddrWithPageBits, size);
 
@@ -744,12 +771,12 @@ uint16_t FRAM9::getSize()
 //
 //  FRAM9  PROTECTED
 //
-void FRAM9::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM9::_writeBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   // Device uses Address Pages
-  uint8_t DeviceAddrWithPageBits = _address | ((memaddr & 0x0100) >> 8);
+  uint8_t DeviceAddrWithPageBits = _address | ((memAddr & 0x0100) >> 8);
   _wire->beginTransmission(DeviceAddrWithPageBits);
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr & 0xFF));
 
   uint8_t * p = obj;
   for (uint8_t i = size; i > 0; i--)
@@ -760,12 +787,12 @@ void FRAM9::_writeBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
 }
 
 
-void FRAM9::_readBlock(uint16_t memaddr, uint8_t * obj, uint8_t size)
+void FRAM9::_readBlock(uint16_t memAddr, uint8_t * obj, uint8_t size)
 {
   // Device uses Address Pages
-  uint8_t DeviceAddrWithPageBits = _address | ((memaddr & 0x0100) >> 8);
+  uint8_t DeviceAddrWithPageBits = _address | ((memAddr & 0x0100) >> 8);
   _wire->beginTransmission(DeviceAddrWithPageBits);
-  _wire->write((uint8_t) (memaddr & 0xFF));
+  _wire->write((uint8_t) (memAddr & 0xFF));
   _wire->endTransmission();
   _wire->requestFrom(DeviceAddrWithPageBits, size);