+++ libraries/FRAM_I2C/CHANGELOG.md
@@ -5,12 +5,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.4] - 2023-09-04
+- merge #39 add **read64(), write64()**  (kudos to pvogt09)
+- merge #40 add **MB85RC64V** info (kudos to pvogt09)
+- rewrite **getManufacturerID, getProductID, getSize()**
+  to handle non existent metadata (See #38)
+- update keywords.txt
+- update unit test
+- improved readability parameter names
+- update readme.md
+
+
 ## [0.5.3] - 2023-04-28
 - major redo FRAM9 and FRAM11, see #35
 - update FRAM32 to match FRAM
 - update readme.md.
 - minor edits
-
 
 ## [0.5.2] - 2023-04-22
 - add **FRAM_ML** class - see FRAM_MULTILANGUAGE.md